requirements.txt
# GA4 & Search Console Analytics Platform
# Professional Quarterly Report Generator for Nonprofits
# =====================================================

# Google APIs
google-analytics-data==0.18.5
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Export Formats
openpyxl==3.1.2
XlsxWriter==3.1.9
python-pptx==0.6.23
reportlab==4.0.7
Pillow==10.1.0

# Dashboard
streamlit==1.29.0
plotly==5.18.0

# Utilities
python-dateutil==2.8.2
pyyaml==6.0.1
rich==13.7.0
click==8.1.7
python-dotenv==1.0.0

# Caching
diskcache==5.6.3

